+++ classes/privacy/provider.php
@@ -26,9 +26,6 @@ namespace local_twittercard\privacy;
 
 defined('MOODLE_INTERNAL') || die();
 
-// Backward compat.
-interface_exists('\core_privacy\local\metadata\null_provider') || die();
-
 /**
  * Privacy Subsystem implementation for local_twittercard implementing null_provider.
  *
@@ -40,22 +37,13 @@ class provider implements
     // This plugin does not store any personal user data.
     \core_privacy\local\metadata\null_provider {
 
-    // 3.4+ runs only on PHP 7.0+ so 'return type declaration' is supported and required,
-    // due to the way the privacy null_provider interface has been implemented in 3.4+.
-    // A polyfill - /privacy/classes/local/legacy_polyfill.php - comes to help here in keeping the same plugin implementation,
-    // regardless the Moodle version, 3.3 vs 3.4+.
-    // Ref.: https://docs.moodle.org/dev/Privacy_API#Difference_between_Moodle_3.3_and_more_recent_versions.
-
-    // This is the trait to be included to actually benefit from the polyfill.
-    use \core_privacy\local\legacy_polyfill;
-
     /**
      * Get the language string identifier with the component's language
      * file to explain why this plugin stores no data.
      *
      * @return  string
      */
-    public static function _get_reason() {
+    public static function get_reason(): string {
         return 'privacy:metadata';
     }
 }